+++ application/GUI/Downloads/DownloadsTreeModel.h
@@ -126,7 +126,7 @@ namespace GUI
       );
 
       Tree* root;
-      QHash<int, Tree*> indexedEntries; // Entries received from the Core. They all have an ID.
+      QHash<quint64, Tree*> indexedEntries; // Entries received from the Core. They all have an ID.
    };
 
    bool operator>(const Protos::GUI::State::Download& d1, const Protos::GUI::State::Download& d2);
